requirements.txt
alabaster==0.7.12
amqp==5.0.6
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
astroid==2.9.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
black==20.8b1
celery==5.1.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorlog==5.0.1
coverage==5.5
cryptography==36.0.0
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==3.2
django-admin-sortable2==1.0
django-allauth==0.44.0
django-celery-beat==2.2.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.2
django-filter==2.4.0
django-grappelli==2.15.1
django-model-utils==4.1.1
django-redis==4.12.1
django-request-logging==0.7.3
django-stubs==1.7.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docutils==0.18.1
drf-jwt==1.19.1
easy-thumbnails==2.7.1
factory-boy==3.2.0
Faker==10.0.0
filelock==3.4.0
flake8==3.9.0
flake8-isort==4.0.0
identify==2.4.0
idna==3.3
imagesize==1.3.0
inflect==5.3.0
iniconfig==1.1.1
ipdb==0.13.7
ipython==7.30.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
kombu==5.2.2
lazy-object-proxy==1.6.0
livereload==2.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.4.0
pluggy==0.13.1
pre-commit==2.11.1
prompt-toolkit==3.0.24
psycopg2==2.8.6
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==3.0.6
pytest==6.2.2
pytest-django==4.1.0
pytest-sugar==0.9.4
python-crontab==2.6.0
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
redis==3.5.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==3.5.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
termcolor==1.1.0
testfixtures==6.18.3
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.4.3
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.13.3
